chore: moved mar computation to utils
